Week9/day5/requirements.txt
pyautogen==0.2.35
openai
python-dotenv
faiss-cpu
numpy